Use a Factory for Singleton Instances (#18)
By using dependency injection we had the problem that use of the services a lot of classes needed to be instantiated in order to pass them to the service constructors. 
With adding a Factor, we can just call it to get the instance of the appropriate service without to take care of the dependencies. 
For the current services a singleton is not an issue as it doesn't have a state or references. So the factory returns a singleton for performance optimisations.