Fix Factory Class
The factory class had a major issue as it stored singleton objects of needed classes as attributes. But these attributes were not initaliased before using as dependency.
Now a completely new approach is taken to get the Singleton instances of the service.
Each service has its own method inside of the factory and the dependencies are injected by calling those methods so it is guaranteed that an isntance exists.